Program can extract features from a new document (ie, list of hashtags) and return a dictionary of hashtags that are present (value = True) and that are not present (value = False).
In main(), extract hashtags from lib tweets to a list of liberal hashtags. Do the same for conservative tweets. A variable called 'classifier' adds a label to the list and creates a tuple. HASHTAGS is a list of these tuples.

By feeding the output of each function into the next function:
get_unique_hashtags -> get_hashtags_as_features -> extract_features returns a dictionary of 'hits' or matches in a given document.